Make an '--export-tasks' option
This causes the build to fail, and dumps a huge amount of output, but
the resulting script does seem to work, kinda.

This is also horrible currently because the server doesn't quit when you
CTRL+C. Probably because it expects to be running lots of idle callbacks
and instead it just runs a huge mainloop-blocking function.

Need to decide if the dump_task code needs to be so slow. If so, need to
split it up into small tasks so the mainloop doesn't get blocked.
Or, hopefully it can just be made really fast.